requirements.txt
PyYAML>=6.0
Pillow>=10.0
torch>=2.1
diffusers>=0.30
transformers>=4.40
accelerate>=0.30
safetensors>=0.4
